+++ composer.json
@@ -22,8 +22,8 @@
 	"require": {
 		"php": ">=8.1",
 		"cakephp/cakephp": "^5.0.0",
-		"dereuromark/cakephp-tools": "dev-cake5",
-		"dereuromark/cakephp-tinyauth": "dev-cake5"
+		"dereuromark/cakephp-tools": "^3.0.0",
+		"dereuromark/cakephp-tinyauth": "^4.0.0"
 	},
 	"require-dev": {
 		"phpunit/phpunit": "^10.2",
@@ -55,7 +55,7 @@
 		"lowest-setup": "composer update --prefer-lowest --prefer-stable --prefer-dist --no-interaction && cp composer.json composer.backup && composer require --dev dereuromark/composer-prefer-lowest && mv composer.backup composer.json"
 	},
 	"prefer-stable": true,
-	"minimum-stability": "stable",
+	"minimum-stability": "RC",
 	"config": {
 		"allow-plugins": {
 			"dealerdirect/phpcodesniffer-composer-installer": true
