+++ src/TinyAuthBackendPlugin.php
@@ -2,8 +2,11 @@
 
 namespace TinyAuthBackend;
 
+use Cake\Console\CommandCollection;
 use Cake\Core\BasePlugin;
 use Cake\Routing\RouteBuilder;
+use TinyAuthBackend\Command\ImportCommand;
+use TinyAuthBackend\Command\InitCommand;
 
 /**
  * Plugin for TinyAuthBackend
@@ -27,6 +30,19 @@ class TinyAuthBackendPlugin extends BasePlugin {
 				},
 			);
 		});
+	}
+
+	/**
+	 * Console hook
+	 *
+	 * @param \Cake\Console\CommandCollection $commands The command collection
+	 * @return \Cake\Console\CommandCollection
+	 */
+	public function console(CommandCollection $commands): CommandCollection {
+		$commands->add('tiny_auth_backend init', InitCommand::class);
+		$commands->add('tiny_auth_backend import', ImportCommand::class);
+
+		return $commands;
 	}
 
 }
